refactor: Applied new `current_path` member variable to `Command` class.
This new attribute is set to the `fs::current_path()` output, which is
the full path to the current directory the user is on. Given that
many commands refer to the current path, it is ideal to introduce this
attribute to limit the function calls to get said path to just once.

Additionally, a small refactor to the `cat` function was made.